Moving importMembersToTerm from term to import service
